requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.6
django-celery-beat==2.5.0
django-crispy-forms==1.14.0
django-environ==0.11.2
django-formtools==2.4.1
django-phonenumber-field==7.2.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
kombu==5.3.2
packaging==23.2
phonenumberslite==8.13.23
Pillow==10.1.0
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
redis==5.0.1
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
whitenoise==6.6.0
